# mobile/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MerchantHub AI",
    "description": "Merchant Engagement Platform with AI-powered review analysis",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/reviews/business/{business_id}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "List Business Reviews",
        "description": "List active reviews for a business.\n\n**Path:** business_id\n**Response:** Reviews with AI analysis, replies, and photo URLs",
        "operationId": "list_business_reviews_api_v1_reviews_business__business_id__get",
        "parameters": [
          {
            "name": "business_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Business Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewResponse"
                  },
                  "title": "Response List Business Reviews Api V1 Reviews Business  Business Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/reported": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "List Reported Reviews",
        "description": "Admin: list reviews flagged for moderation.",
        "operationId": "list_reported_reviews_api_v1_reviews_reported_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewResponse"
                  },
                  "type": "array",
                  "title": "Response List Reported Reviews Api V1 Reviews Reported Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/reviews/admin/all": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "List Admin Reviews",
        "description": "Admin: browse reviews across every business and status, optionally scoped\nto one business (drives both the \"All reviews\" browse view and a\nbusiness drill-down's review history).\n\n**Query:** business_id (optional scope), page (default 1), page_size (default 20, cap 100)\n**Response:** Reviews of every status, each carrying its business summary",
        "operationId": "list_admin_reviews_api_v1_reviews_admin_all_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "business_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "nullable": true,
              "title": "Business Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewResponse"
                  },
                  "title": "Response List Admin Reviews Api V1 Reviews Admin All Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews": {
      "post": {
        "tags": [
          "Reviews"
        ],
        "summary": "Create Review",
        "description": "Submit a review \u2014 triggers automatic AI text analysis.\n\n**Request:** business_id, rating (1-5), title, body (min 10 chars)\n**Response:** Review with AI analysis attached",
        "operationId": "create_review_api_v1_reviews_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/reviews/{review_id}": {
      "patch": {
        "tags": [
          "Reviews"
        ],
        "summary": "Update Review",
        "description": "Edit own review. Re-runs AI analysis if body changes.",
        "operationId": "update_review_api_v1_reviews__review_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Reviews"
        ],
        "summary": "Delete Review",
        "description": "Delete own review.",
        "operationId": "delete_review_api_v1_reviews__review_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{review_id}/like": {
      "post": {
        "tags": [
          "Reviews"
        ],
        "summary": "Like Review",
        "description": "Like a review (idempotent).",
        "operationId": "like_review_api_v1_reviews__review_id__like_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{review_id}/report": {
      "post": {
        "tags": [
          "Reviews"
        ],
        "summary": "Report Review",
        "description": "Report inappropriate review.",
        "operationId": "report_review_api_v1_reviews__review_id__report_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewReportCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{review_id}/reply": {
      "post": {
        "tags": [
          "Reviews"
        ],
        "summary": "Reply To Review",
        "description": "Merchant responds to a review.",
        "operationId": "reply_to_review_api_v1_reviews__review_id__reply_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{review_id}/moderate": {
      "post": {
        "tags": [
          "Reviews"
        ],
        "summary": "Moderate Review",
        "description": "Admin: hide or restore a review. action=hide|restore|remove",
        "operationId": "moderate_review_api_v1_reviews__review_id__moderate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ai/reviews/{review_id}": {
      "get": {
        "tags": [
          "AI Analysis"
        ],
        "summary": "Get Review Analysis",
        "description": "Get AI analysis for a specific review.\n\n**Path:** review_id\n**Response:** Sentiment, summary, positives, complaints, suggested response",
        "operationId": "get_review_analysis_api_v1_ai_reviews__review_id__get",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewCreate": {
        "properties": {
          "business_id": {
            "type": "string",
            "format": "uuid",
            "title": "Business Id"
          },
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating"
          },
          "title": {
            "type": "string",
            "nullable": true,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "minLength": 10,
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "business_id",
          "rating",
          "body"
        ],
        "title": "ReviewCreate"
      },
...
      "ReviewResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "business_id": {
            "type": "string",
            "format": "uuid",
            "title": "Business Id"
          },
          "author_id": {
            "type": "string",
            "format": "uuid",
            "title": "Author Id"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "title": {
            "type": "string",
            "nullable": true,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "like_count": {
            "type": "integer",
            "title": "Like Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "author": {
            "$ref": "#/components/schemas/UserResponse",
            "nullable": true
          },
          "ai_analysis": {
            "$ref": "#/components/schemas/AIAnalysisResponse",
            "nullable": true
          },
          "reply": {
            "$ref": "#/components/schemas/ReplyResponse",
            "nullable": true
          },
          "photo_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Photo Urls",
            "default": []
          },
          "business": {
            "$ref": "#/components/schemas/BusinessSummary",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "business_id",
          "author_id",
          "rating",
          "body",
          "status",
          "like_count",
          "created_at"
        ],
        "title": "ReviewResponse"
      },
...
      "ReviewUpdate": {
        "properties": {
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "nullable": true,
            "title": "Rating"
          },
          "title": {
            "type": "string",
            "nullable": true,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "minLength": 10,
            "nullable": true,
            "title": "Body"
          }
        },
        "type": "object",
        "title": "ReviewUpdate"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}